Generacion dinamica de cards de cursos en la vista
