+++ src/componentes/Grid/ItemStyles.css
@@ -1,5 +1,3 @@
 .card-item {
-    width: 100%;
     box-shadow: none !important;
-  }
-  
+}
